requirements.txt
streamlit==1.36.0
youtube-transcript-api==0.6.2
google-generativeai==0.7.2
python-dotenv==1.0.1
fpdf==1.7.2